Cambiar promedio a float64

+++ tp1/tp1.go
@@ -29,7 +29,7 @@ func (p Productos) CalcularPrecios(ids ...int) []Carrito {
 
 // Promedio recibe el id de un producto y retorna el precio promedio
 // de ese producto usando los precios de todos los supermercados.
-func (p Productos) Promedio(idProducto int) int {
+func (p Productos) Promedio(idProducto int) float64 {
 	return 0
 }
